Tableau Visualization
The project is a dashboard created using Tableau software.

- The dashboard provides an overview of COVID-19 case and death rates across the world.
- The dashboard is divided into several sections, including a world map, a search bar, and filters.
- The World Map section displays the COVID-19 case and death rates for each country on a map, allowing users to visually explore the spread of the pandemic.
- The Search Bar allows users to search for specific countries or regions to filter the data on the map.
- The Filters section provides options to filter the data by date range and rate type.
- The dashboard is designed to be interactive, allowing users to filter and drill down into the data to gain deeper insights.

Overall, the dashboard provides a comprehensive view of COVID-19 case and death rates across the world, allowing users to easily explore the impact of the pandemic in different countries and regions. This information can be used to make informed decisions about travel, public health policy, and other related issues.